:art: [#1193] -- re-organize file-deletion related code into single module
